feat: Debounce useChange on useTree hook
